Update vdr-skinflatplus (latest git)
Current VDR has changed locking mechanism which is incompatible with
present version of skinflatplus. One of the symptoms of this is that
only current and past EPG events are displayed in Schedule menu. There
may be other issues which I didn't spot.
Hence update package to latest commit which has proper adjustments.